Re-arrange the string module
